more descriptive commit message:
moved commands unused in RobotContainer to autoscommands directory
