+++ src/main/java/frc/robot/commands/Teleop/BringIn.java
@@ -36,16 +36,12 @@ public class BringIn extends Command {
     if (!shooterDone) {
       shooter.setHingePower(-0.05);
       if (shooter.switchPressed()) {
-        shooter.setHingePower(0.0);
-        shooter.setEncoderPos(Constants.ShooterConstants.DOWN_LIMIT);
         shooterDone = true;
       }
     }
     if (!intakeDone) {
       intake.setHingePower(-0.05);
       if (intake.switchPressed()) {
-        intake.setHingePower(0.0);
-        intake.setEncoderPos(Constants.IntakeConstants.UP_LIMIT);
         intakeDone = true;
       }
     }
