+++ src/main/java/frc/robot/commands/Teleop/IntakeDown.java
@@ -2,17 +2,19 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.Autoscommands;
+// NOTE this is not being used this maybe remove
+package frc.robot.commands.Teleop;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.Intake;
 import frc.robot.subsystems.PivotIntake;
 
-public class IntakeUp extends Command {
+public class IntakeDown extends Command {
+  private static final String Pivotintake = null;
   private PivotIntake intake;
   public boolean done = false;
-  /** Creates a new IntakeUp. */
-  public IntakeUp(PivotIntake intake) {
+
+  /** Creates a new IntakeDown. */
+  public IntakeDown(PivotIntake intake) {
     this.intake = intake;
     // Use addRequirements() here to declare subsystem dependencies.
     addRequirements(intake);
@@ -21,7 +23,7 @@ public class IntakeUp extends Command {
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
-    intake.liftIntake();
+    intake.dropIntake();
     done = true;
   }
