+++ src/main/java/frc/robot/commands/Teleop/IntakeUp.java
@@ -2,19 +2,17 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-// NOTE this is not being used this maybe remove
-package frc.robot.commands.Autoscommands;
+package frc.robot.commands.Teleop;
 
 import edu.wpi.first.wpilibj2.command.Command;
+import frc.robot.subsystems.Intake;
 import frc.robot.subsystems.PivotIntake;
 
-public class IntakeDown extends Command {
-  private static final String Pivotintake = null;
+public class IntakeUp extends Command {
   private PivotIntake intake;
   public boolean done = false;
-
-  /** Creates a new IntakeDown. */
-  public IntakeDown(PivotIntake intake) {
+  /** Creates a new IntakeUp. */
+  public IntakeUp(PivotIntake intake) {
     this.intake = intake;
     // Use addRequirements() here to declare subsystem dependencies.
     addRequirements(intake);
@@ -23,7 +21,7 @@ public class IntakeDown extends Command {
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
-    intake.dropIntake();
+    intake.liftIntake();
     done = true;
   }
